PlatformerMechanicDashInProgress
Dash attempted and works when standing still. however when moving it works once and never again. i believe it either has something to do with decelerating too fast or the clamping getting in the way but my quick fixes don't seem to work right now so i'm gonna move on